Hash Tables chaining with Singly Linked Lists
This is a C Program to implement hash tables using linked list. Hash table will have ‘n’ number of buckets. 
To insert a node into the hash table, we need to find the hash index for the given key. And it could be calculated using the hash function. Example: hashIndex = key % noOfBuckets 
Move to the bucket corresponds to the above-calculated hash index and insert the new node at the end of the list. To delete a node from the hash table, get the key from the user, calculate the hash index, move to the bucket corresponds to the calculated hash index, search the list in the current bucket to find and remove the node with the given key. Finally, remove the node with given key, if it is present.